Move module definitions to separate file.
